requirements.txt
discord.py==2.3.2
python-dotenv>=1.0
pyyaml>=6.0
jinja2>=3.1
openai>=1.0
redis>=5.0
google-api-python-client>=2.0
beautifulsoup4>=4.12
requests>=2.31
